Moving tests under public folder
