+++ sdk/communication/communication-chat/test/public/chatClient.mocked.spec.ts
@@ -3,9 +3,9 @@
 
 import sinon from "sinon";
 import { assert } from "chai";
-import { ChatClient, CreateChatThreadRequest } from "../src";
-import * as RestModel from "../src/generated/src/models";
-import { apiVersion } from "../src/generated/src/models/parameters";
+import { ChatClient, CreateChatThreadRequest } from "../../src";
+import * as RestModel from "../../src/generated/src/models";
+import { apiVersion } from "../../src/generated/src/models/parameters";
 import { baseUri, generateToken } from "./utils/connectionUtils";
 import {
   AzureCommunicationTokenCredential,
