+++ sdk/communication/communication-chat/test/public/chatClient.spec.ts
@@ -3,7 +3,7 @@
 
 import { isLiveMode, Recorder } from "@azure/test-utils-recorder";
 import { assert } from "chai";
-import { ChatClient, ChatThreadClient } from "../src";
+import { ChatClient, ChatThreadClient } from "../../src";
 import { createTestUser, createRecorder, createChatClient } from "./utils/recordedClient";
 import { isNode } from "@azure/core-http";
 import sinon from "sinon";
