+++ sdk/communication/communication-chat/test/public/chatThreadClient.spec.ts
@@ -3,7 +3,7 @@
 
 import { Recorder } from "@azure/test-utils-recorder";
 import { assert } from "chai";
-import { ChatClient, ChatThreadClient } from "../src";
+import { ChatClient, ChatThreadClient } from "../../src";
 import { createTestUser, createRecorder, createChatClient } from "./utils/recordedClient";
 import { CommunicationIdentifier, getIdentifierKind } from "@azure/communication-common";
 import { Context } from "mocha";
