+++ sdk/communication/communication-chat/test/public/utils/connectionUtils.ts
